Add support for extra-sync - extra branches to sync with.
I've got a repo of text files that I edit from many machines.

Each machine has its own branch that only it edits. For that branch,
the usual sync operation should always succeed.

I also try to keep origin/master up-to-date with whichever branch was
edited most recently. Sometimes this fails - usually because I've made
conflicting edits on two machines while they were offline, and so
manual intervention is needed to resolve the conflict.

This change adds this ability into git-sync. It looks for a list of
extra branches to sync with in branch.$branch_name.extra-sync. If this
is found, the same sync logic is used for each of those in turn, after
the main upstream branch is synced